Big spike!
* Moved different components to separate files.
* Added Help popover.
* Upgraded to React 0.10.
* Added a needed polyfill and included es5-shim/sham
* Upgraded jQuery and added source maps.
* Added a different loading indicator (popover near Artie).
* Added scorecard nav with dropdowns.

As I hack on this, I realize there's a bunch of optimizations that can
be had if we do them server-side, things we couldn't do when in the
Ember-Data straightjacket. For example, we could load the scorecards
already grouped and sorted, saving time on the client-side to group
them. Same with the matrix -- it'll be much easier to load and render
if it's already grouped appropriately. We can also decrease the
request URL size if we use query parameters for the related object
instead of the ginormous list of ids.