Fixed generating scorecard/platform report only.
The scorecard and platform reports need the test_results coverage files.
However, to save some time, they check to see if those files already
exist in the temp directory. If they do, the test result html generation
functions are not called.

Those functions added the riak_lib ebin's to the code path, but the
scorecard/platform functions did not. This meant that if the test results
were generated, and then the scorecards generated, they would not be able
to find the modules covered, resulting in truncated coverage report html.